requirements.txt
requests
Pillow
beautifulsoup4
lxml
scikit-learn
scipy
numpy
PySide6
PyInstaller
